requirements.txt
flask

# Para Variaveis de Ambiente
python-dotenv

# Usamos o 'psycopg2-binary' pois é mais fácil de instalar
psycopg2-binary

# ORM (Mapeamento Objeto-Relacional) para facilitar a interação com o DB
Flask-SQLAlchemy

# Para gerenciar migrações do banco de dados
Flask-Migrate

# Para segurança de senhas (vem com o Flask, mas é bom listar)
Werkzeug